# Conflicts:
#	app/Livewire/Forms/ExtendDataset.php
#	app/Livewire/Forms/UploadDataset.php
#	resources/views/livewire/forms/extend-dataset.blade.php
#	resources/views/livewire/forms/upload-dataset.blade.php

+++ app/ActionRequestService/Handlers/ExtendDatasetHandler.php
@@ -57,7 +57,11 @@ class ExtendDatasetHandler extends BaseHandler
                 // If class exists in parent, update annotation to parent class
                 if ($childClass && isset($parentClasses[$childClass->name])) {
                     $annotation->annotation_class_id = $parentClasses[$childClass->name]->id;
-                    $annotationsClassesToUpdate[] = $annotation->toArray();
+                    $annotationArray = $annotation->toArray();
+                    if (isset($annotationArray['segmentation']) && is_array($annotationArray['segmentation'])) {
+                        $annotationArray['segmentation'] = json_encode($annotationArray['segmentation']);
+                    }
+                    $annotationsClassesToUpdate[] = $annotationArray;
                     continue;
                 }
                 // If new class, add to parent dataset
@@ -142,7 +146,9 @@ class ExtendDatasetHandler extends BaseHandler
     }
     public function errorResponse(string $errorMessage, ActionRequest $request = null): mixed
     {
-        $request->delete();
+        if($request){
+            $request->delete();
+        }
         return ['type' => 'error', 'message' => $errorMessage];
     }
 }
